Flag compatibility w/Simplified CommonJS Wrapper
min-require does not work with standard Universal Module Definition
loaders as those include a check for define.amd which is not defined
for this implementation.

In addition, UMD loaders use the AMD version of define that expects a
list of dependencies and the factory method, which is not implemented
either.

One possible workaround is to add a flag to define to indicate the UMD
loader that it uses a SCJSW signature and then load the module
accordingly.

I.e: the UMD loader commonjsStrict.js example should be updated adding
the following condition:

if (typeof define === 'function' && define.cjs) {
    define("commonJsStrict", function (require, module, exports) {
        factory(exports, require('b'));
    });
}